facial-emotion-marks/requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

